chore: fixing build
