+++ src/app/course/cbt/submit/page.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React from "react"
 import {
   Box,
