+++ src/app/course/leaderboard/page.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React from "react"
 import {
   Box,
