+++ src/app/courses/page.tsx
@@ -13,22 +13,13 @@ import Footer from "@/components/Footer"
 import CoursesSearch from "@/components/courses/CoursesSearch"
 
 import CourseData from "@/components/courses/CourseData"
+import { coursefilter } from "@/utils/util"
 
 export type FilterItem = {
   label: string
   value: string
 }
 
-export const coursefilter: FilterItem[] = [
-  { label: "All Courses", value: "all" },
-  { label: "Web Development", value: "web" },
-  { label: "Mobile Development", value: "mobile" },
-  { label: "Data Science", value: "data science" },
-  { label: "UI/UX Design", value: "ui/ux design" },
-  { label: "Digital Marketing", value: "digital marketing" },
-  { label: "Business", value: "business" },
-  { label: "Photography", value: "photography" },
-]
 // export const metadata = generateMetadata({
 //   title: "Horace Learning Management Solution | Horace Courses",
 //   description:
