+++ src/app/dashboard/courses/[id]/page.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { Box } from "@mui/material"
 import { useParams } from "next/navigation"
