+++ src/utils/util.ts
@@ -1,0 +1,12 @@
+import { FilterItem } from "@/app/courses/page"
+
+export const coursefilter: FilterItem[] = [
+  { label: "All Courses", value: "all" },
+  { label: "Web Development", value: "web" },
+  { label: "Mobile Development", value: "mobile" },
+  { label: "Data Science", value: "data science" },
+  { label: "UI/UX Design", value: "ui/ux design" },
+  { label: "Digital Marketing", value: "digital marketing" },
+  { label: "Business", value: "business" },
+  { label: "Photography", value: "photography" },
+]
